Add EPT Zstandard loader.
