+++ pointclouds/lion_takanawa_ept_zst/ept-hierarchy/0-0-0-0.json
@@ -1,0 +1,11 @@
+{
+  "0-0-0-0": 35535,
+  "1-0-0-0": 29751,
+  "1-0-0-1": 24986,
+  "1-0-1-0": 57518,
+  "1-0-1-1": 33380,
+  "1-1-0-0": 25549,
+  "1-1-1-0": 42454,
+  "2-1-1-1": 50549,
+  "2-1-1-2": 42267
+}
