+++ src/PointCloudEptGeometry.js
@@ -68,10 +68,19 @@ export class PointCloudEptGeometry {
 
 		let hierarchyType = info.hierarchyType || 'json';
 
-		let dataType = info.dataType || 'laszip';
-		this.loader = dataType == 'binary'
-			? new Potree.EptBinaryLoader()
-			: new Potree.EptLaszipLoader();
+        const dataType = info.dataType;
+        if (dataType == 'laszip') {
+            this.loader = new Potree.EptLaszipLoader();
+        }
+        else if (dataType == 'binary') {
+			this.loader = new Potree.EptBinaryLoader();
+        }
+        else if (dataType == 'zstandard') {
+            this.loader = new Potree.EptZstandardLoader();
+        }
+        else {
+            throw new Error('Could not read data type: ' + dataType);
+        }
 	}
 };
