+++ src/loader/ept/ZstandardLoader.js
@@ -1,0 +1,12 @@
+import {EptBinaryLoader} from "./BinaryLoader.js";
+
+export class EptZstandardLoader extends EptBinaryLoader {
+    extension() {
+        return '.zst';
+    }
+
+    workerPath() {
+        return Potree.scriptPath + '/workers/EptZstandardDecoderWorker.js';
+    }
+};
+
